refactor(se-danskebank-bankid): double to bigdecimal and cleanup

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/danskebank/fetchers/investment/rpc/GroupAccountEntity.java
@@ -66,11 +66,11 @@ public class GroupAccountEntity {
 
     public Portfolio toTinkPortfolio(ListSecuritiesResponse response) {
         Portfolio portfolio = new Portfolio();
-        portfolio.setTotalValue(response.getMarketValue());
+        portfolio.setTotalValue(response.getMarketValue().doubleValue());
         portfolio.setRawType(type);
         portfolio.setType(getTinkPortfolioType());
         portfolio.setUniqueIdentifier(accountIdentifier);
-        portfolio.setTotalProfit(response.getPerformance());
+        portfolio.setTotalProfit(response.getPerformance().doubleValue());
         return portfolio;
     }
